moving components into a dedicated subfolder
