+++ src/components/comp2.vue
