+++ src/components/comp3.vue
