+++ src/components/comp6.vue
