+++ src/index.js
@@ -1,12 +1,12 @@
 import Vue from 'vue'
 
-import comp1 from '@/comp1'
-import comp2 from '@/comp2'
-import comp3 from '@/comp3'
-import comp4 from '@/comp4'
-import comp5 from '@/comp5'
-import comp6 from '@/comp6'
-import comp7 from '@/comp7'
+import comp1 from '@/components/comp1'
+import comp2 from '@/components/comp2'
+import comp3 from '@/components/comp3'
+import comp4 from '@/components/comp4'
+import comp5 from '@/components/comp5'
+import comp6 from '@/components/comp6'
+import comp7 from '@/components/comp7'
 
 export const mainApp = new Vue({
   el: '#main-app',
