+++ test/unit/specs/components/comp3.spec.js
@@ -1,5 +1,5 @@
 import { shallow } from 'vue-test-utils'
-import comp3 from '@/comp3'
+import comp3 from '@/components/comp3'
 
 describe('comp3', () => {
   test('should match snapshot', (done) => {
