+++ test/unit/specs/components/comp7.spec.js
@@ -1,6 +1,6 @@
 import { mount } from 'vue-test-utils'
-import comp7 from '@/comp7.vue'
-import todoItem from '@/todoItem.vue'
+import comp7 from '@/components/comp7.vue'
+import todoItem from '@/components/todoItem.vue'
 
 describe('comp7', () => {
   let component, vm, element
